refactor(cli-core): move handle_build into leo_cli_core::commands::build
`handle_build` (and its 8 helpers: `compile_leo_source_directory`,
`parse_leo_source_directory`, `parse_leo_source_directory_library`,
`build_leo_source_directory_library`, `validate_compiled_programs`,
`validate_compiled_programs_inner`, `write_interface_abis`,
`ensure_parent_dir`, `remove_legacy_build_artifacts`) moves out of
`crates/leo/src/cli/commands/build.rs` and into
`crates/leo-cli-core/src/commands/build.rs`. The two `DisassembleProcess` /
`ProgramForValidation` helper types come with it.

Also moves:
- `crates/leo/src/errors.rs` → `crates/leo-cli-core/src/errors.rs`,
  with every `pub(crate)` flipped to `pub`. `crates/leo` re-exports as
  `pub(crate) use leo_cli_core::errors;` so existing
  `crate::errors::*` callsites keep working.
- `LOCAL_PROGRAM_DEFAULT_EDITION` and `format_program_size` from
  `crates/leo/src/cli/commands/common/util.rs` into
  `crates/leo-cli-core::commands` so the shared `handle_build` can
  reach them without depending on the CLI binary.

The `LeoBuild` clap struct in `crates/leo` is now a thin wrapper: it
resolves `--network` / `--endpoint` defaults and calls
`leo_cli_core::commands::build::handle_build(&self.options, network,
&endpoint, ..., &package_path, &home_path)`.

`leo-cli-core` Cargo.toml gains `leo-abi` and `itertools` (used by
the moved code). Snarkvm/tracing/ureq were already there.
