+++ crates/leo-cli-core/src/lib.rs
@@ -43,8 +43,11 @@ pub mod options;
 pub mod project;
 
 // Native-only modules — disk + HTTP + snarkVM-umbrella helpers lifted out
-// of `crates/leo-package` so that crate stays purely wasm-buildable. Each
-// module is `#[cfg(not(target_arch = "wasm32"))]`-gated internally.
+// of `crates/leo-package` and `crates/leo` so those crates stay purely
+// wasm-buildable / clap-only. Each module is
+// `#[cfg(not(target_arch = "wasm32"))]`-gated internally.
+pub mod commands;
+pub mod errors;
 pub mod network;
 pub mod package_fetch;
 pub mod package_init;
